+++ src/main/java/sarangbang/site/challengemember/repository/ChallengeMemberRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public interface ChallengeMemberRepository extends JpaRepository<ChallengeMember, Long> {
 
-    Long countByChallengeId(Long challengeId);
+    int countByChallengeId(Long challengeId);
 
     Optional<ChallengeMember> findChallengeMemberByUser_IdAndChallenge_Id(String userId, Long challengeId);
     List<ChallengeMember> findByChallengeId(Long challengeId);
